move injection code into separate file
